internal/controller/spa/spa.go
/*
 * This file is part of xrDebug.
 *
 * (c) Rodolfo Berrios <rodolfo@chevere.org>
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */

package spa

import (
	"net/http"
)

// Handle creates an HTTP handler that serves a Single Page Application (SPA) from HTML content.
// serves contents with the appropriate content type and status codes.
func Handle(content []byte) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if r.URL.Path != "/" {
			http.NotFound(w, r)
			return
		}
		w.Header().Set("Content-Type", "text/html")
		w.Header().Set("Content-Encoding", "gzip")
		w.Write(content)
	}
}
